uplift fuse to index
