+++ components/PlantSearch.js
@@ -1,19 +1,4 @@
-import { useState, useMemo } from "react";
-import Fuse from "fuse.js";
-
-export default function SearchBar({ plants }) {
-  const [query, setQuery] = useState("");
-
-  const searchPlants = useMemo(() => {
-    if (!query) return plants;
-    const fuseInstance = new Fuse(plants, {
-      keys: ["name", "botanicalName"],
-      threshold: 0.5,
-    });
-    const results = fuseInstance.search(query);
-    return results.map((result) => result.item);
-  }, [query, plants]);
-
+export default function SearchBar({ query, setQuery }) {
   return (
     <div>
       <input
@@ -22,13 +7,6 @@ export default function SearchBar({ plants }) {
         value={query}
         onChange={(e) => setQuery(e.target.value)}
       />
-      <ul>
-        {searchPlants.map((item) => (
-          <li key={item.id}>
-            {item.name} — {item.botanicalName}
-          </li>
-        ))}
-      </ul>
     </div>
   );
 }
